解决订单号重复的问题

+++ src/main/java/run/halo/starter/QuartzService.java
@@ -13,11 +13,16 @@ import run.halo.app.extension.ExtensionClient;
 import run.halo.app.extension.ListOptions;
 import run.halo.app.extension.index.query.Query;
 import java.util.List;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+
 
 import static run.halo.app.extension.index.query.QueryFactory.and;
 import static run.halo.app.extension.index.query.QueryFactory.equal;
 import static run.halo.app.extension.index.query.QueryFactory.lessThan;
 import static run.halo.app.extension.index.query.QueryFactory.or;
+
 
 @Component
 @RequiredArgsConstructor
@@ -27,30 +32,48 @@ public class QuartzService {
     private final ExtensionClient client;
     @PostConstruct
     public void init() {
-        CronUtil.schedule("*/3 * * * * *", new Task() {
-            @Override
-            public void execute() {
-                try {
-                    Query query = and(equal("spec.payStatus", PayStatus.PENDING.name()),lessThan("spec.expireTime", String.valueOf(System.currentTimeMillis())));
+        //延迟30秒启动
+        // 创建一个 ScheduledExecutorService 实例
+        ScheduledExecutorService executorService = Executors.newScheduledThreadPool(1);
 
-                    ListOptions options = ListOptions.builder().fieldQuery(query).build();
+        // 延迟30秒执行定时任务
+        executorService.schedule(() -> {
 
-                    List<PaymentRecord> records = client.listAll(PaymentRecord.class, options, Sort.by(
-                        Sort.Order.asc("spec.orderId")) );
-                    Flux.fromIterable(records)
-                        .flatMap(record -> Mono.fromRunnable(() -> client.delete(record)))
-                        .collectList()
-                        .subscribe();
-                } catch (Exception e) {
-                    e.printStackTrace();
+            CronUtil.schedule("*/3 * * * * *", new Task() {
+                @Override
+                public void execute() {
+                    try {
+                        Query query = and(equal("spec.payStatus", PayStatus.PENDING.name()),
+                            lessThan("spec.expireTime",
+                                String.valueOf(System.currentTimeMillis())));
+
+                        ListOptions options = ListOptions.builder().fieldQuery(query).build();
+
+                        List<PaymentRecord> records =
+                            client.listAll(PaymentRecord.class, options, Sort.by(
+                                Sort.Order.asc("spec.orderId")));
+                        if (records.isEmpty()) {
+                            System.out.println("没有找到过期的订单");
+                            return;
+                        } else {
+                            System.out.println("找到过期的订单{}个" + records.size());
+                            Flux.fromIterable(records)
+                                .flatMap(record -> Mono.fromRunnable(() -> client.delete(record)))
+                                .collectList()
+                                .subscribe();
+                        }
+
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
+
                 }
+            });
 
-            }
-        });
-
-        // 支持秒级别定时任务
-        CronUtil.setMatchSecond(true);
-        CronUtil.start();
+            // 支持秒级别定时任务
+            CronUtil.setMatchSecond(true);
+            CronUtil.start();
+        }, 30, TimeUnit.SECONDS);
     }
 
 }
